requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Environment and config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
supabase==2.9.0
postgrest==0.17.2

# Google services
google-generativeai==0.8.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# HTTP client
httpx==0.27.2

# Utilities
pytz==2024.2
python-dateutil==2.9.0.post0

# Image processing (optional)
Pillow==11.0.0
